dashboard/requirements.txt
Dashboard
python-qpid-proton>=0.30
python-dotenv>=0.13.0
